requirements.txt
Appium-Python-Client==5.1.1
selenium==4.32.0
easyocr==1.7.2
pillow==11.0.0
opencv-python-headless==4.11.0.86
numpy==1.26.4
requests==2.32.3

#pip freeze > requirements.txt
# source venv/bin/activate
#